test: add some tests
