+++ .github/workflows/main.yml
@@ -26,5 +26,8 @@ jobs:
       - name: Run lint
         run: npm run lint
 
+      - name: Run test
+        run: npm run test
+
       - name: Run build
         run: npm run build
