+++ .gitignore
@@ -1,4 +1,5 @@
 .idea
 node_modules
+coverage
 .rts*
 dist
