--- .release-it.json
@@ -1,6 +1,0 @@
-{
-  "hooks": {
-    "before:init": ["npm run lint"],
-    "after:bump": "npm run build"
-  }
-}
