+++ package.json
@@ -12,9 +12,9 @@
   "scripts": {
     "build": "microbundle --jsx React.createElement --external react ./src/index.tsx --output dist",
     "start": "microbundle --jsx React.createElement --external react ./src/index.tsx --output dist --watch",
+    "test": "jest",
     "typecheck": "tsc --noEmit",
-    "lint": "eslint --fix \"src/**/*.{js,jsx,ts,tsx}\"",
-    "release": "release-it"
+    "lint": "eslint --fix \"src/**/*.{js,jsx,ts,tsx}\""
   },
   "repository": {
     "type": "git",
@@ -38,11 +38,16 @@
   },
   "devDependencies": {
     "@semantic-release/git": "^9.0.0",
+    "@testing-library/jest-dom": "^5.11.9",
+    "@testing-library/react": "^11.2.5",
+    "@testing-library/react-hooks": "^5.0.3",
+    "@types/jest": "^26.0.20",
     "@types/node": "^14.14.25",
     "@types/react": "^17.0.1",
     "@types/react-dom": "^17.0.0",
     "@typescript-eslint/eslint-plugin": "^4.14.2",
     "@typescript-eslint/parser": "^4.14.2",
+    "babel-eslint": "^10.1.0",
     "eslint": "^7.19.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-airbnb-typescript": "^12.3.1",
@@ -52,7 +57,11 @@
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-react-hooks": "^4.2.0",
+    "jest": "^26.6.3",
     "microbundle": "^0.13.0",
-    "semantic-release": "^17.3.7"
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "semantic-release": "^17.3.7",
+    "ts-jest": "^26.5.0"
   }
 }
