+++ src/react-app-env.d.ts
@@ -1,6 +1,6 @@
 /// <reference types="node" />
 /// <reference types="react" />
-/// <reference types="react-dom" />
+/// <reference types="react-dom" />s
 
 declare namespace NodeJS {
   interface ProcessEnv {
